fix: convert Price property to decimal

+++ PetShop_Upgrade/DTOS/Discounts/DiscountDTO.cs
@@ -8,9 +8,9 @@ namespace PetShop_Upgrade.DTOS.Discounts
         public string Code { get; set; }
         public string DiscountName { get; set; }
         public string? Description { get; set; }
-        public double DiscountValue { get; set; } // Giá trị Giảm giá
-        public double? MinOrderValue { get; set; } // Giá trị thấp nhất có thể áp dụng
-        public double? MaxDiscountAmount { get; set; } // Giá trị giảm giá cao nhất
+        public decimal DiscountValue { get; set; } // Giá trị Giảm giá
+        public decimal? MinOrderValue { get; set; } // Giá trị thấp nhất có thể áp dụng
+        public decimal? MaxDiscountAmount { get; set; } // Giá trị giảm giá cao nhất
         public int? MaxUsage { get; set; } // Tổng số lần mã này được nhập
         public int? MaxUsagePerUser { get; set; } // Số lần tối đa một người dùng được phép nhập mã này
         public DiscountType DiscountType { get; set; } = DiscountType.FIXED_AMOUNT;// 0: percentage, 1: fixed amount
