+++ PetShop_Upgrade/DTOS/Products/Admin/AdminProductFilterDTO.cs
@@ -7,8 +7,8 @@ namespace PetShop_Upgrade.DTOS.Products.Admin
         public int? ProductId { get; set; }
         public string? ProductName { get; set; }
         public int? CategoryId { get; set; }
-        public double? ImportPrice { get; set; }
-        public double? SellingPrice { get; set; }
+        public decimal? ImportPrice { get; set; }
+        public decimal? SellingPrice { get; set; }
         public ProductType? Type { get; set; } // 0: Pet, 1: Toy, 2: Food
     }
 }
