+++ PetShop_Upgrade/DTOS/Products/Admin/UpdateProductDTO.cs
@@ -7,8 +7,8 @@ namespace PetShop_Upgrade.DTOS.Products.Admin
     {
         public string ProductName { get; set; }
         public string? Description { get; set; }
-        public double ImportPrice { get; set; }
-        public double SellingPrice { get; set; }
+        public decimal ImportPrice { get; set; }
+        public decimal SellingPrice { get; set; }
         public int MainImageIndex { get; set; } = 0;
         public int CategoryId { get; set; }
         public IsActive IsActive { get; set; }
