+++ PetShop_Upgrade/DTOS/Products/Client/ProductResponseRequestDTO.cs
@@ -15,8 +15,8 @@ namespace PetShop_Upgrade.DTOS.Products.Client
         public int Id { get; set; }
         public string ProductName { get; set; }
         public string? Description { get; set; }
-        public double ImportPrice { get; set; }
-        public double SellingPrice { get; set; }
+        public decimal ImportPrice { get; set; }
+        public decimal SellingPrice { get; set; }
         public ProductType Type { get; set; }
         public IsActive IsActive { get; set; }
         public string? CategoryName { get; set; }
